requirements.txt
streamlit>=1.32.0
langchain>=0.2.0
langchain-community>=0.2.0
langchain-openai>=0.1.6
openai>=1.30.0
faiss-cpu>=1.8.0
gitpython>=3.1.43
python-dotenv>=1.0.1
tiktoken>=0.6.0
pygments>=2.17.2